Use at own risk!
This is a generic updater. It tries to keep any changes that you have made including databases. If it fails (because you have edited the same lines in the files that have also been changed by this repository) then you can always just run the updater with the argument --hard or simply -h to force a hard reset of the files. The hard reset will first backup your files to "/home/pi/thermos_backup/year_month_date-hour_minute"